Add useCreateUser hook for creating a new user
This commit adds a new hook called useCreateUser, which is responsible for handling the creation of a new user. It uses the Apollo Client to send a mutation to the GraphQL API, specifically the createUser mutation. The hook takes in an input object containing the user's email and password, and returns the created user object, including the user's ID and email.

The createUser mutation is defined in the gql template literal and takes in the createUserInput object as a parameter. The mutation returns the ID and email of the created user.

Additionally, two new files were added: AbstractModel.ts and User.ts. AbstractModel.ts defines an interface called AbstractModel, which includes a property for the ID. User.ts imports the AbstractModel interface and extends it to include an email property, defining the User interface.

These changes are necessary to implement the functionality for creating a new user in the application.